requirements.txt
flask
gevent
python-dotenv
emoji
mutagen
aiohttp